Suggest version upgrades when old version is missing
In cases where library_versions does not contain a record for the
current version of a library that is depended on, we do not recommend
any upgrades. This happens because we do a join with `library_versions`
to find the sort key of the current version.

We have enough information to construct the sort key without a join, so
this PR fixes that.

This was discovered while investigating why, when most of Flow's projects
depend on `com.typesafe.play.sbt-plugin` 2.8.5, and Dependency knows
about 2.8.11, the upgrade isn't recommended.